+++ snowddl/parser/function.py
@@ -1,7 +1,9 @@
 from snowddl.blueprint import (
     FunctionBlueprint,
+    AccountObjectIdent,
     Ident,
     SchemaObjectIdentWithArgs,
+    ArgumentWithType,
     NameWithType,
     DataType,
     StageWithPath,
@@ -19,8 +21,26 @@ function_json_schema = {
         },
         "arguments": {
             "type": "object",
-            "additionalProperties": {
-                "type": "string"
+            "minItems": 1,
+            "items": {
+                "anyOf": [
+                    {
+                        "type": "string"
+                    },
+                    {
+                        "type": "object",
+                        "properties": {
+                            "type": {
+                                "type": "string"
+                            },
+                            "default": {
+                                "type": "string"
+                            },
+                        },
+                        "required": ["type"],
+                        "additionalProperties": False
+                    }
+                ]
             }
         },
         "returns": {
@@ -81,6 +101,19 @@ function_json_schema = {
         "handler": {
             "type": "string"
         },
+        "external_access_integrations": {
+            "type": "array",
+            "items": {
+                "type": "string"
+            },
+            "minItems": 1
+        },
+        "secrets": {
+            "type": "object",
+            "additionalProperties": {
+                "type": "string"
+            },
+        },
         "comment": {
             "type": "string"
         }
@@ -103,31 +136,7 @@ class FunctionParser(AbstractParser):
         self.parse_schema_object_files("function", function_json_schema, self.process_function)
 
     def process_function(self, f: ParsedFile):
-        arguments = [NameWithType(name=Ident(k), type=DataType(t)) for k, t in f.params.get("arguments", {}).items()]
-
-        if isinstance(f.params["returns"], dict):
-            # Returns table
-            returns = [NameWithType(name=Ident(k), type=DataType(t)) for k, t in f.params["returns"].items()]
-        else:
-            # Returns value
-            returns = DataType(f.params["returns"])
-
-        if f.params.get("imports"):
-            imports = [
-                StageWithPath(
-                    stage_name=build_schema_object_ident(self.env_prefix, i["stage"], f.database, f.schema),
-                    path=self.normalise_stage_path(i["path"]),
-                )
-                for i in f.params.get("imports")
-            ]
-        else:
-            imports = None
-
-        if f.params.get("packages"):
-            packages = f.params.get("packages")
-        else:
-            packages = None
-
+        arguments = self.get_arguments(f)
         base_name = self.validate_name_with_args(f.path, arguments)
 
         bp = FunctionBlueprint(
@@ -137,16 +146,67 @@ class FunctionParser(AbstractParser):
             language=f.params.get("language", "SQL"),
             body=f.params.get("body"),
             arguments=arguments,
-            returns=returns,
+            returns=self.get_returns(f),
             is_secure=f.params.get("is_secure", False),
             is_strict=f.params.get("is_strict", False),
             is_immutable=f.params.get("is_immutable", False),
             is_memoizable=f.params.get("is_memoizable", False),
             runtime_version=str(f.params.get("runtime_version")) if f.params.get("runtime_version") else None,
-            imports=imports,
-            packages=packages,
+            imports=self.get_imports(f),
+            packages=f.params.get("packages"),
             handler=f.params.get("handler"),
+            external_access_integrations=self.get_external_access_integrations(f),
+            secrets=self.get_secrets(f),
             comment=f.params.get("comment"),
         )
 
         self.config.add_blueprint(bp)
+
+    def get_arguments(self, f: ParsedFile):
+        arguments = []
+
+        for arg_name, arg in f.params.get("arguments", {}).items():
+            # Short syntax
+            if isinstance(arg, str):
+                arg = {"type": arg}
+
+            arguments.append(
+                ArgumentWithType(
+                    name=Ident(arg_name),
+                    type=DataType(arg["type"]),
+                    default=str(arg["default"]) if "default" in arg else None,
+                )
+            )
+
+        return arguments
+
+    def get_returns(self, f: ParsedFile):
+        if isinstance(f.params["returns"], dict):
+            # Returns table
+            return [NameWithType(name=Ident(k), type=DataType(t)) for k, t in f.params["returns"].items()]
+
+        return DataType(f.params["returns"])
+
+    def get_imports(self, f: ParsedFile):
+        if f.params.get("imports"):
+            return [
+                StageWithPath(
+                    stage_name=build_schema_object_ident(self.env_prefix, i["stage"], f.database, f.schema),
+                    path=self.normalise_stage_path(i["path"]),
+                )
+                for i in f.params.get("imports")
+            ]
+
+        return None
+
+    def get_external_access_integrations(self, f: ParsedFile):
+        if f.params.get("external_access_integrations"):
+            return [AccountObjectIdent(self.env_prefix, i) for i in f.params.get("external_access_integrations")]
+
+        return None
+
+    def get_secrets(self, f: ParsedFile):
+        if f.params.get("secrets"):
+            return {k: build_schema_object_ident(self.env_prefix, v, f.database, f.schema) for k, v in f.params.get("secrets").items()}
+
+        return None