+++ snowddl/resolver/_utils.py
@@ -29,7 +29,9 @@ def compare_dynamic_param_value(bp_value: Union[bool, int, float, str], existing
 
 
 def dtypes_from_arguments(arguments: str) -> str:
+    arguments = arguments.translate(str.maketrans('', '', '[] '))
+
     start_dtypes_idx = arguments.index("(")
     finish_dtypes_idx = arguments.index(")")
 
-    return ",".join([a.strip(" ") for a in arguments[start_dtypes_idx + 1 : finish_dtypes_idx].split(",")])
+    return ",".join(a for a in arguments[start_dtypes_idx + 1 : finish_dtypes_idx].split(","))