+++ snowddl/resolver/table.py
@@ -36,6 +36,10 @@ class TableResolver(AbstractSchemaObjectResolver):
             if r["is_external"] == "Y":
                 continue
 
+            # Skip event tables
+            if r["is_event"] == "Y":
+                continue
+
             full_name = f"{r['database_name']}.{r['schema_name']}.{r['name']}"
             existing_objects[full_name] = {
                 "database": r["database_name"],