+++ snowddl/settings.py
@@ -18,6 +18,7 @@ class SnowDDLSettings(BaseModelWithConfig):
     refresh_user_passwords: bool = False
     refresh_future_grants: bool = False
     refresh_stage_encryption: bool = False
+    refresh_secrets: bool = False
     clone_table: bool = False
     exclude_object_types: List[ObjectType] = []
     include_object_types: List[ObjectType] = []